Reboot on exits now that bootloader catches most issues
Hanging on exit used to be really useful when outputing to the HDMI port
(the current default) so that you could capture errors. Now that
bootloader is around, it catches application issues for those not
updating to use a serial port or running over Ethernet. Switching this
behavior is useful for the hardware regression tests to keep them from
hanging.