Reduce fwup redundancy by using an EEx template
The final fwup.conf file is still being committed to allow time to grow
trust in using a template language to generate the fwup.conf.

This refactor is really needed for TRYBOOT support due to significantly
more fwup.conf redundancy to migrate pre-TRYBOOT devices.

The mbr definitions were moved to the fwup-common.conf to align this
with other Raspberry Pis.